Updated CSS styling and links in documents pages
Updated styles across the different year sections in the 'documents-all/page.tsx' with a new 'dashed' css class to emphasize links. Additionally, separated the links for photos and videos into a new list for better user navigation. These changes enhance the page design and improve user interaction with the display of documents.