Add randomTimeString to API calls; Adjust banner document link
Introduced a dynamic element to the API calls in route.js by adding randomTimeString in the query parameters. This is done to prevent caching of API responses and ensure that the latest data is fetched everytime.

Also, corrected the document link related to the ISESER2018 banner by adjusting the extension from .png to .pdf for accurate file access. This change ensures that the document can be properly displayed and downloaded.