Optimize document sections with year parameter and link adjustments
Introduced 'DocumentWithYear' function in 'page.tsx' to reduce redundancy by allowing passing of years as parameter. Hyphens were added in the urls replacing underscores to conform with naming convention. In the 'sitemap.xml/route.js' file, certain lines relating to 'ABSTRACT BOOK' and 'BANNER.pdf' were removed or modified. This was done in a bid to streamline the document section and improve user experience.