"Restrict Abstract Book widget display for 2018"
Updated client.tsx to display 'Abstract Book' widget only for year '2018'. Previously, the widget was being displayed for all years where 'paper_abst_page' data was available. This change was necessary to align with the conference's decision to only provide the abstract book for 2018 event. It ensures that users don't see misleading content for other years.