Refactor description string formatting in page.tsx
Updated the formatting of the description string in the page.tsx file. Change was made to improve the clarity and readability of the description string generated. Now, the generated description string reads as '[en_title] - Abstract of an article titled' as opposed to the previous way of formatting: 'This is an abstract of an article titled [en_title]'.